+++ tests/plugin/http/sw_websockets/services/__init__.py
@@ -14,31 +14,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import requests
-import websockets
-
-import asyncio
-
-if __name__ == '__main__':
-    from fastapi import FastAPI
-    import uvicorn
-
-    app = FastAPI()
-
-    @app.get('/users')
-    async def application():
-        res = requests.get('http://provider:9091/users', timeout=5)
-        websocket_pong = await websocket_ping()
-        return {'http': res.json(), 'websocket': websocket_pong}
-
-    async def websocket_ping():
-        async with websockets.connect('ws://provider:9091/ws', extra_headers=None) as websocket:
-            await websocket.send('Ping')
-
-            response = await websocket.recv()
-            await asyncio.sleep(0.5)
-
-            await websocket.close()
-            return response
-
-    uvicorn.run(app, host='0.0.0.0', port=9090)