+++ tests/plugin/http/sw_websockets/services/provider.py
@@ -3,7 +3,7 @@
 # contributor license agreements.  See the NOTICE file distributed with
 # this work for additional information regarding copyright ownership.
 # The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
+# (the 'License'); you may not use this file except in compliance with
 # the License.  You may obtain a copy of the License at
 #
 #     http://www.apache.org/licenses/LICENSE-2.0
@@ -14,7 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-import time
 
 from starlette.websockets import WebSocketDisconnect
 
@@ -23,12 +22,6 @@ if __name__ == '__main__':
     import uvicorn
 
     app = FastAPI()
-
-
-    @app.get('/users')
-    async def application():
-        time.sleep(0.5)
-        return {'song': 'Despacito', 'artist': 'Luis Fonsi'}
 
 
     @app.websocket('/ws')
@@ -45,5 +38,6 @@ if __name__ == '__main__':
         except WebSocketDisconnect:
             ...
         return 'ok'
+
 
     uvicorn.run(app, host='0.0.0.0', port=9091)